requirements.txt
adafruit-circuitpython-neopixel
gpiozero
rpi-ws281x
